Fixed CropTileConfig properties to float values instead of int values in CroTileConfig (issue #28)

+++ source/core/src/main/com/csse3200/game/entities/configs/CropTileConfig.java
@@ -5,7 +5,7 @@ package com.csse3200.game.entities.configs;
  */
 public class CropTileConfig extends BaseEntityConfig {
 
-    public int initialSoilQuality = 1;  //Default value for initial soil quality
-    public int initialWaterContent = 1; //Default value for initial water content
+    public float initialSoilQuality = 1;  //Default value for initial soil quality
+    public float initialWaterContent = 1; //Default value for initial water content
 
 }
